added logout button to topbar
