+++ src/main/cliapp/src2/layout/index.js
@@ -1,0 +1,3 @@
+import { SiteLayout } from "./SiteLayout";
+
+export { SiteLayout };
